backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23
sqlmodel>=0.0.16
python-dotenv>=1.0.0
pydantic>=2.0
openai>=1.0.0
httpx>=0.27.0
redis>=5.0.0
